# config/openapi.yml
openapi: 3.0.3
info:
  title: YouGov API example
  description: YouGov API example declaration
  termsOfService: http://swagger.io/terms/
  contact:
    email: leonardo.bruno@outlook.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 1.0.0
...
servers:
  - url: /api/v1
...
paths:
  /teams/{team_name}:
    get:
      description: Returns a specific team
      operationId: getByName
      security:
        - bearerAuth: []
      parameters:
        - name: team_name
          required: true
          in: path
          description: The name of the team
          schema:
            type: string
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetTeamByName"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    GetTeamByName:
      type: object